+++ aldrin/src/lib.rs
@@ -88,7 +88,9 @@ pub use channel::{
     UnclaimedReceiver, UnclaimedSender,
 };
 pub use client::Client;
-pub use discoverer::{Discoverer, DiscovererBuilder, DiscovererEventKind, DiscovererEventRef};
+pub use discoverer::{
+    Discoverer, DiscovererBuilder, DiscovererEvent, DiscovererEventKind, DiscovererEventRef,
+};
 pub use error::Error;
 pub use handle::Handle;
 pub use lifetime::{Lifetime, LifetimeId, LifetimeScope};